sql interface refactoring
